refactor(taxes): extract SuggestModal component
Move SuggestModal to src/pages/taxes/components/SuggestModal.tsx.
Remove unused Account import from Taxes.tsx.
Taxes.tsx 767→706L.

Part of #181
